fix bug in grid_fit / barycentric coordinates
problem when sigma is small, being 10 sigma form the edges can be true even if you are on the edge, and then barycentric_coordinates throws off an error because you try to put a weight on a point that isn't on the grid.
fixed by forcing grid_fit to return L large enough that you are both 10 sigma and 10 grid points from the edges.